// scanapp-js/src/html5-qrcode/ui/scanner/qr-scan-dashboard.d.ts
import { ScanTypeSelector } from "./scan-type-selector";
import { Html5QrcodeScanType } from "../../core";
import { Html5QrcodeHandler } from "../../html5-qrcode-handler";
import { PersistedDataManager } from "../../storage";
import { OnFileSelected } from "../../file-scan/file-scan-core";
import { CameraDevice } from "../../camera/core";
type Runnable = () => void;
export declare class QrScanDashboard {
    private readonly scanTypeSelector;
    private readonly persistedDataManager;
    private readonly html5QrcodeHandler;
    private readonly defaultZoomValueIfSupported;
    private readonly onFileSelectedCallback;
    private readonly cameraQueryHandler;
    private readonly container;
    private readonly childContainer;
    private topDashboard;
    private middleDashboard;
    private bottomDashboard;
    private cameraScanImage;
    private fileScanImage;
    private requestPermissionButton;
    private fileScanUi;
    constructor(scanTypeSelector: ScanTypeSelector, persistedDataManager: PersistedDataManager, html5QrcodeHandler: Html5QrcodeHandler, defaultZoomValueIfSupported: number | undefined, onFileSelectedCallback: OnFileSelected, cameraQueryHandler: Runnable);
    private createChildContainer;
    private createDashboard;
    private createTopDashboard;
    private createMiddleDashboard;
    private createBottomDashboard;
    private setupCleanDashboard;
    private setupNonCameraFileOnlyDash;
    private setupNonCameraDash;
    private setupNonCameraDashWithoutPermissionButton;
    private setupCameraDash;
    private renderCameraScanImage;
    private renderFileScanImage;
    private renderFileScanUi;
    private createPermissionButton;
    enablePermissionButton(): void;
    enableOrSetupPermissionButton(): void;
    renderCamera(cameraList: Array<CameraDevice>): void;
    private renderCameraInternal;
    private renderCameraDashboard;
    optionallySetImageName(fileName: string): void;
    hideMiddleDash(): void;
    showMiddleDash(): void;
    static createAndRender(parentContainer: HTMLDivElement, currentScanType: Html5QrcodeScanType, scanTypeSelector: ScanTypeSelector, persistedDataManager: PersistedDataManager, html5QrcodeHandler: Html5QrcodeHandler, defaultZoomValueIfSupported: number | undefined, onFileSelectedCallback: OnFileSelected, cameraQueryHandler: Runnable): QrScanDashboard;
}
export {};